+++ nuages/src/components/pendingReviewTable.tsx
@@ -1,0 +1,20 @@
+import { Text } from "@chakra-ui/react"
+
+
+interface PendingReviewTableProps {
+    // props
+
+}
+
+
+const PendingReviewTable = (props: PendingReviewTableProps) => {
+
+
+    return (
+        <Text>
+            Pending Review
+        </Text>
+    )
+}
+
+export default PendingReviewTable
