+++ nuages/src/pages/home.tsx
@@ -1,0 +1,12 @@
+
+import styles from "../styles/navigation.module.css";
+
+const Home = () => {
+    return (
+        <div className={styles.cole}>
+            <h1>Home</h1>
+        </div>
+    );
+};
+
+export default Home;
