+++ nuages/src/pages/pharmaciesPage.tsx
